requirements.txt
streamlit
google-cloud-firestore
python-dotenv
firebase
firebase_admin